添加记录信息

+++ assist/python/cg/playlist.py
@@ -6,18 +6,20 @@ import concurrent.futures
 
 from pathlib import Path
 
-
+import operator
 
 import re
 import json
 import sys
+import asyncio
 
 root_path=Path(__file__).parent.parent.resolve()
 sys.path.append(str(root_path ))
 sys.path.append( os.path.join(root_path,'base') )
 
-from base import exception_decorator,logger_helper,except_stack,normal_path,fetch_sync,decrypt_aes_128,get_folder_path
-from base import download_sync,move_file,get_homepage_url,is_http_or_https,hash_text,delete_directory,merge_video
+from base import exception_decorator,logger_helper,except_stack,normal_path,fetch_sync,decrypt_aes_128,get_folder_path,UpdateTimeType
+from base import download_async,download_sync,move_file,get_homepage_url,is_http_or_https,hash_text,delete_directory,merge_video
+from base import as_normal,MultiThreadCoroutine
 
 class video_info:
     def __init__(self,url) -> None:
@@ -109,6 +111,13 @@ def get_playlist(url):
 
     return playlist
     
+    
+
+    
+
+        
+
+
 @exception_decorator()
 def handle_playlist(url_list,temp_paths,key,iv):
     if not url_list or not temp_paths:
@@ -116,35 +125,38 @@ def handle_playlist(url_list,temp_paths,key,iv):
     
 
     logger= logger_helper("下载文件",Path(temp_paths[0]).parent)
-    success=True
-    
+
+    logger.trace("开始")
     def decode():
         if not key or not iv:
             return None 
         def _decode(encrypted_data):
             return decrypt_aes_128(key,iv,encrypted_data)
         return _decode
-        
-        
     
-    # 使用线程池并行下载
-    with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
-        futures = [executor.submit(download_sync, url, temp_path,decode())   for  url,temp_path in zip(url_list,temp_paths)]
+    async def download(semaphore,args:list|tuple):
+        async with semaphore:
+            url, temp_path=args
+            result=  await download_async(url, temp_path,decode()) 
+            return result
         
-        # 等待所有任务完成
-        for future in concurrent.futures.as_completed(futures):
-            try:
-                if not future.result():
-                    success=False
-            except Exception as e:
-                logger.error("下载异常",except_stack())
-                success=False
+    multi_thread_coroutine = MultiThreadCoroutine(download,list(zip(url_list,temp_paths)))
+    try:
+        asyncio.run(multi_thread_coroutine.run_tasks()) 
+        success=multi_thread_coroutine.success
+        if not success:
+            info=[multi_thread_coroutine.fail_infos,except_stack()]
+            info_str="\n".join(info)
+            logger.error("异常",f"\n{info_str}\n",update_time_type=UpdateTimeType.ALL)
+        return multi_thread_coroutine.success
+    except Exception as e:
+        logger.error("下载异常",except_stack(),update_time_type=UpdateTimeType.ALL)
+        return False
+    
 
-    
-    return success
             
 def temp_paths(count,temp_dir):
-    return [normal_path(os.path.join(temp_dir, f"{index+1}.mp4"))    for index in range(count)]
+    return [normal_path(os.path.join(temp_dir, f"{index}.mp4"))    for index in range(count)]
 
 
 def decryp_video(org_path,dest_path,key,iv):
@@ -192,8 +204,8 @@ def main(url,url_pre,dest_name):
     temp_path=normal_path(os.path.join(temp_dir,f"{dest_hash}.mp4")) 
     dest_path=normal_path(os.path.join(root_path,f"{dest_name}.mp4"))
     
-    
-
+    logger= logger_helper("下载",f"{dest_name}-{dest_hash}")
+    logger.info("开始")
         
     os.makedirs(temp_dir, exist_ok=True)
     
@@ -206,8 +218,9 @@ def main(url,url_pre,dest_name):
     info_list=info.playlist
     if not info_list:
         return
-    with open(os.path.join(temp_dir,"url.txt"),"w",encoding="utf-8-sig") as f:
-        json.dump(info_list,f,ensure_ascii=False,indent=4)
+    with open(os.path.join(root_path,"urls",f"{dest_name}-{dest_hash}.json"),"w",encoding="utf-8-sig") as f:
+        info={"url":url,"name":dest_name,"hash":dest_hash, "playlist":info_list}
+        json.dump(info,f,ensure_ascii=False,indent=4)
     
 
     if not url_pre:
@@ -215,14 +228,25 @@ def main(url,url_pre,dest_name):
         
     url_list=[get_real_url(urls[2],url)  for urls in info_list]
     temp_path_list=temp_paths(len(url_list),temp_dir)
-    
 
-    
-    
-    
     success=handle_playlist(url_list,temp_path_list,key,iv)
     if not success:
-        return
+        already_path_list=[]
+        losts=[]
+        
+        
+        
+        for item in temp_path_list:
+            if os.path.exists(item):
+                already_path_list.append(item)
+            else:
+                losts.append(item)
+        with open(os.path.join(root_path,"urls",f"{dest_name}-{dest_hash}-lost.json"),"w",encoding="utf-8-sig") as f:
+            json.dump(losts,f,ensure_ascii=False,indent=4)
+            
+        temp_path_list=already_path_list
+        if not temp_path_list:
+            return
 
     # decryp_video(org_path,dest_path,key,iv)
     
@@ -231,9 +255,11 @@ def main(url,url_pre,dest_name):
     merge_video(temp_path_list,temp_path)
     move_file(temp_path,dest_path)
     
-    delete_directory(temp_dir)
-    
-    
+    if success:
+        delete_directory(temp_dir)
+        logger.info("完成" ,update_time_type=UpdateTimeType.ALL)
+    else:
+        logger.error("部分缺失",update_time_type=UpdateTimeType.ALL)
     
 def get_key(url):
 
@@ -242,28 +268,24 @@ def get_key(url):
     
 if __name__=="__main__":
     
+
     
 
     
     lst=[
-        # ("https://vip.lz-cdn.com/20220917/33091_abc5295d/1200k/hls/mixed.m3u8","https://vip.lz-cdn.com/20220917/33091_abc5295d/1200k/hls/","3D肉蒲团1"),
-        # ("https://ikcdn01.ikzybf.com/20240422/zaL0A0dC/2000kb/hls/index.m3u8","https://kkzycdn.com:65/20240422/zaL0A0dC/2000kb/hls/","3D肉蒲团"),
-        # ("https://v8.tlkqc.com/wjv8/202310/09/6FnT1gEfrp1/video/1000k_720/hls/index.m3u8","","金瓶梅2008"),
-        # ("https://ukzyll.ukubf6.com/20220526/9ROENkuT/2000kb/hls/index.m3u8","https://ukzyll.ukubf6.com","金瓶梅2"),
-        # ("https://hd.ijycnd.com/play/negMER9b/index.m3u8","","金瓶梅II爱的奴隶"),
-        # ("https://hd.ijycnd.com/play/Le32QD9d/index.m3u8","","隔壁的呻吟声"),
-        # ("https://ikcdn01.ikzybf.com/20221201/6iQAY1nx/2000kb/hls/index.m3u8","","波多野结衣之双飞调教"),
-        # ("https://v8.fentvoss.com/sdv8/202311/04/24UGLwGmFz1/video/1000k_720/hls/index.m3u8","","巨乳妻耻辱性活"),
-        # ("https://v8.fentvoss.com/sdv8/202310/20/3Ufvxm9cAP1/video/1000k_720/hls/index.m3u8","","看护妇日记野兽般的午后"),
-        # ("https://s1.bfllvip.com/video/xingjiaoyi2/%E4%B8%AD%E5%AD%97/index.m3u8","","性交易2"),
-        # ("https://v.gsuus.com/play/penK8PEa/index.m3u8","","交换工作_漂亮的女性员工","加密"),
-        ("https://ukzy.ukubf4.com/20220403/Gf23AIB0/2000kb/hls/index.m3u8","https://ukzy.ukubf4.com","四大名捕之入梦妖灵"),
-        # ("https://s1.bfzyll.com/video/nvyuangong2dui2xingai/HD/index.m3u8","","女员工_2对2性爱"),
-        # ("https://v6.fentvoss.com/sdv6/202311/05/SHZXaaR9L91/video/1000k_720/hls/index.m3u8","","全方位性爱观察_性爱观"),
-        # ("https://hd.ijycnd.com/play/QdJEvWlb/index.m3u8","","性爱寄宿家庭_轮流性爱","加密"),
+
+
+        ("https://ukzyll.ukubf6.com/20220530/34sNHPqK/2000kb/hls/index.m3u8","","夏娃"),
+
+
         
         
     ]
+    
+    # names=[print(name[2],hash_text(name[2]))  for name in lst]
+    
+    
+    
     for url,url_pre,dest_name,*args in lst:
         main(url,url_pre,dest_name)
